update testimonial call api render
